UserService save changes

+++ src/main/java/com/example/app/dto/UserDTO.java
@@ -3,6 +3,7 @@ package com.example.app.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
@@ -10,7 +11,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDTO {
-    private Integer id;
+    private long id;
     private String name;
     private Integer age;
     private String email;
